Several fixes for the debugger
 - Status lable in GUI debugger
 - AddBreak max line number is now uint16_t max (16383)
 - IntDecode/IntEncode for the 3 types of integers
 - Removed magic hex numbers and replaced with enum's
 - pydebugger.py cleans up between script installs
 - Deferred eval response
 - Eval can now trigger breakpoints
 - Removed "busy" state concept
 - Fixed ERR_READ when Ctrl-C or close debug client
 - Eval/Putvar submitted with enter key
 - Can set breakpoints with space key
 - Scroll bar for source and console windows
 - Windows python GUI support

+++ console/ajs_console_c.h
@@ -67,9 +67,9 @@ int AJS_ConsoleConnect(AJS_ConsoleCtx* ctx, const char* deviceName, volatile sig
  *
  * @param ctx           Console context
  * @param script        Eval string
- * @return              1 on success, 0 on failure
+ * @return              Result of AJSConsole.Eval()
  */
-int AJS_ConsoleEval(AJS_ConsoleCtx* ctx, const char* script);
+int8_t AJS_ConsoleEval(AJS_ConsoleCtx* ctx, const char* script);
 
 /**
  * Signal the AllJoyn.js target to reboot
